Sidekiq setup
Here I delegate the creation and the api call of each job to a worker.
There are several problems with this, like jobs duplication as worker are asynchronous.
A first solution would be to add uniqueness validation on job table.
An other solution could be to use workers for the api call, and save the record in db before.
Also, I should probably use a worker for the csv importation and store the file on S3, but this is an other thing.